A Simple Susceptible-High-risk-latent-Low-risk-latent-Infected-Recovered Infectious Disease Model

A Simple Susceptible-High-risk-latent-Low-risk-latent-Infected-Recovered Infectious Disease Model

SHLIR_ode(t, x, params)

Arguments

t
The timestep overwhich to calculate derivatives

x
A numeric vector of compartment populations.

params
A named vector of paramter values.

Value

A vector of derivatives

Examples

##Model Input
S_0 <- 989
H_0 <- 10
L_0 <- 0
I_0 <- 1
R_0 <- 0
beta = 3 # Rate of transmission
gamma_H = 1/5 # Rate of progression to active symptoms from high risk latent
nu = 1/2 #Rate of progression from high to low risk latent
gamma_L = 1/100 # Rate of progression to active symptoms for low risk latent
tau = 1/2 # Rate of recovery
dt <- 1

parameters <- c(beta = beta, gamma_H = gamma_H, gamma_L = gamma_L, nu = nu, tau = tau)
inits <- c(S = S_0, H = H_0, L = L_0, I = I_0, R_0 = R_0)

SHLIR_ode(1, inits, parameters)

#> [[1]]
#>      S      S      H      H      I 
#> -2.967 -4.033  5.000  1.500  0.500 
#>