Combine an Infectious Disease Model To a Demographic Model

Combine an Infectious Disease Model To a Demographic Model

combine_to_age_model(df, age_com = 3, compartments = c("S", "S_v", "E",
  "E_v", "I", "I_v"), hold_out_var = c("time", "traj"), total_pop = TRUE)

Arguments

df
A dataframe of model output.

age_com
The number of age compartments that have been modelled.

compartments
A character vector of the disease model compartments to combine.

hold_out_var
The variables that have not been stratified by age.

total_pop
A logical indicating whether to calculate the total population. Defaults to true.

Value

A dataframe which summarises the demographic process of a model.

See also

combine_strat_model_output

Examples

df <- data.frame(S1 = 1, S2 = 1, E1 = 1, E2 = 1, time = 1)

combine_to_age_model(df, age_com = 2, compartments = c("S", "E"), hold_out_var = "time")

#> # A tibble: 1 x 4
#>   age_group_1 age_group_2  time     N
#>         <dbl>       <dbl> <dbl> <dbl>
#> 1           2           2     1     4