Lookup the Details of Parameters Required by a Model

This function simplifies the process of checking which parameters a given idmodelr model depends on. It is effectively an interface to parameter_details via model_details. As fuzzy matching has been used it can also given information of the parameter requirements of a subset of the available models.

required_parameters(model = NULL)

Arguments

model

A character string containing the name of the model of interest. Defaults to NULL.

Value

A dataframe extracted from parameter_details containing the details of the parameters required by the model of interest.

Examples

## Check the parameters required by the "SIR_ode" model
required_parameters("SIR_ode")
#> # A tibble: 2 × 6
#>   parameter parameter_family description                   type  risk_…¹ non_e…²
#>   <chr>     <chr>            <chr>                         <chr> <chr>   <chr>  
#> 1 beta      transmission     Transmission rate = the tran… rate  no      no     
#> 2 tau       recovery         Recovery rate. The reciproca… rate  no      no     
#> # … with abbreviated variable names ¹​risk_stratified, ²​non_exponential

## Use fizzy matching to look at parameters for all SIR models
required_parameters("SIR")
#> # A tibble: 5 × 6
#>   parameter parameter_family description                   type  risk_…¹ non_e…²
#>   <chr>     <chr>            <chr>                         <chr> <chr>   <chr>  
#> 1 beta      transmission     Transmission rate = the tran… rate  no      no     
#> 2 tau       recovery         Recovery rate. The reciproca… rate  no      no     
#> 3 lambda    vaccination      The effectiveness of vaccina… prob… no      no     
#> 4 alpha     vaccination      The coverage of vaccination.  prop… no      no     
#> 5 mu        demographics     The natural mortality rate. … rate  no      no     
#> # … with abbreviated variable names ¹​risk_stratified, ²​non_exponential