Summarise Model Trajectories

This function checks, cleans, aggregates and summarises inputed model trajectories. This function is based on plotTraj from the fitR package written by Sebaustian Funk.

summarise_model(traj = NULL, state.names = NULL, data = NULL,
  time.column = "time", summary = TRUE, replicate.column = "replicate",
  non.extinct = NULL, init.date = NULL, aggregate_to = aggregate_to,
  compartments = compartments, strat = strat, hold_out_var = hold_out_var,
  new_var = new_var, total_pop = total_pop, verbose = FALSE)

Arguments

traj
Data frame, generic model output although tailored to work with pomp model objects.

state.names
A character vector. Names of the state variables to plot. If NULL (default) all state variables are plotted.

data
A data frame. Observation times and observed data. The time column must match that used in the model object, whilt the observation name should also match one of those found in the model object.

time.column
A character vector. The column in the data that indicates time

summary
Logical. If TRUE, the mean, median as well as the 50th and 95th percentile of the trajectories are plotted (default). If FALSE, all individual trajectories are plotted.

replicate.column
A character vector. The column in the data that indicates the replicate (if muliple replicates are to be plotted, i.e. if summary is FALSE

non.extinct
A character vector. Names of the infected states which must be non-zero so the epidemic is still ongoing. When the names of these states are provided, the extinction probability is plotted by computing the proportion of faded-out epidemics over time. An epidemic has faded-out when all the infected states (whose names are provided) are equal to 0. This is only relevant for stochastic models.

init.date
Character. Date of the first point of the time series (default to NULL). If provided, the x-axis will be in calendar format. NB: currently only works if the unit of time is the day.

aggregate_to
A character vector or list specifying the aggregation operations to perfrom on the model output. Operations are carried out in the order specified. Implemented options are; disease, demographic, and incidence.

compartments
A character vector or list specifying the unique compartments to aggregate. May either be specified once for all aggregation functions or for each function seperately.

strat
The number of stratified groups in the model.

hold_out_var
A character vector or list specifying the unique compartments that will not be aggregated. May either be specified once for all aggregation functions or for each function seperately. If compartments is set then this arguement does not need to be used.

new_var
A character vevtor specifying the new variable to add when aggregating incidence.

total_pop
A logical vector (defaults to TRUE) indicating if the total population should be calculated when summarsing the model demographics.

verbose
Logical (defaults to FALSE), indicates if internal process messages should be printed

Value

A list of summarised/aggregated dataframes for an inputed model trajectory. These are the model trajectories, probability of extinction, and model summary trajectories.

See also

plot_model aggregate_model

Examples