Sum a Stratified Variable

Sum a Stratified Variable

summarise_strat_var(df, vars, strat, new_var = "sum")

Arguments

df
A dataframe of model output.

vars
A character vector containing the unstratified variables to summarise

strat
The number of stratifications present in the data set

new_var
The name of the summarised variable

Value

Returns the original dataframe with an additional summarised variable

See also

summarise_var_by_strat

Examples

library(dplyr)
df <- mutate(iris, Petal.Length1 = Petal.Length, Petal.Length2 = Petal.Length)

summarise_strat_var(df, vars = c("Petal.Length"), strat = 2, new_var = "sum")

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species Petal.Length1
#> 1            5.1         3.5          1.4         0.2     setosa           1.4
#> 2            4.9         3.0          1.4         0.2     setosa           1.4
#> 3            4.7         3.2          1.3         0.2     setosa           1.3
#> 4            4.6         3.1          1.5         0.2     setosa           1.5
#> 5            5.0         3.6          1.4         0.2     setosa           1.4
#> 6            5.4         3.9          1.7         0.4     setosa           1.7
#> 7            4.6         3.4          1.4         0.3     setosa           1.4
#> 8            5.0         3.4          1.5         0.2     setosa           1.5
#> 9            4.4         2.9          1.4         0.2     setosa           1.4
#> 10           4.9         3.1          1.5         0.1     setosa           1.5
#> 11           5.4         3.7          1.5         0.2     setosa           1.5
#> 12           4.8         3.4          1.6         0.2     setosa           1.6
#> 13           4.8         3.0          1.4         0.1     setosa           1.4
#> 14           4.3         3.0          1.1         0.1     setosa           1.1
#> 15           5.8         4.0          1.2         0.2     setosa           1.2
#> 16           5.7         4.4          1.5         0.4     setosa           1.5
#> 17           5.4         3.9          1.3         0.4     setosa           1.3
#> 18           5.1         3.5          1.4         0.3     setosa           1.4
#> 19           5.7         3.8          1.7         0.3     setosa           1.7
#> 20           5.1         3.8          1.5         0.3     setosa           1.5
#> 21           5.4         3.4          1.7         0.2     setosa           1.7
#> 22           5.1         3.7          1.5         0.4     setosa           1.5
#> 23           4.6         3.6          1.0         0.2     setosa           1.0
#> 24           5.1         3.3          1.7         0.5     setosa           1.7
#> 25           4.8         3.4          1.9         0.2     setosa           1.9
#> 26           5.0         3.0          1.6         0.2     setosa           1.6
#> 27           5.0         3.4          1.6         0.4     setosa           1.6
#> 28           5.2         3.5          1.5         0.2     setosa           1.5
#> 29           5.2         3.4          1.4         0.2     setosa           1.4
#> 30           4.7         3.2          1.6         0.2     setosa           1.6
#> 31           4.8         3.1          1.6         0.2     setosa           1.6
#> 32           5.4         3.4          1.5         0.4     setosa           1.5
#> 33           5.2         4.1          1.5         0.1     setosa           1.5
#> 34           5.5         4.2          1.4         0.2     setosa           1.4
#> 35           4.9         3.1          1.5         0.2     setosa           1.5
#> 36           5.0         3.2          1.2         0.2     setosa           1.2
#> 37           5.5         3.5          1.3         0.2     setosa           1.3
#> 38           4.9         3.6          1.4         0.1     setosa           1.4
#> 39           4.4         3.0          1.3         0.2     setosa           1.3
#> 40           5.1         3.4          1.5         0.2     setosa           1.5
#> 41           5.0         3.5          1.3         0.3     setosa           1.3
#> 42           4.5         2.3          1.3         0.3     setosa           1.3
#> 43           4.4         3.2          1.3         0.2     setosa           1.3
#> 44           5.0         3.5          1.6         0.6     setosa           1.6
#> 45           5.1         3.8          1.9         0.4     setosa           1.9
#> 46           4.8         3.0          1.4         0.3     setosa           1.4
#> 47           5.1         3.8          1.6         0.2     setosa           1.6
#> 48           4.6         3.2          1.4         0.2     setosa           1.4
#> 49           5.3         3.7          1.5         0.2     setosa           1.5
#> 50           5.0         3.3          1.4         0.2     setosa           1.4
#> 51           7.0         3.2          4.7         1.4 versicolor           4.7
#> 52           6.4         3.2          4.5         1.5 versicolor           4.5
#> 53           6.9         3.1          4.9         1.5 versicolor           4.9
#> 54           5.5         2.3          4.0         1.3 versicolor           4.0
#> 55           6.5         2.8          4.6         1.5 versicolor           4.6
#> 56           5.7         2.8          4.5         1.3 versicolor           4.5
#> 57           6.3         3.3          4.7         1.6 versicolor           4.7
#> 58           4.9         2.4          3.3         1.0 versicolor           3.3
#> 59           6.6         2.9          4.6         1.3 versicolor           4.6
#> 60           5.2         2.7          3.9         1.4 versicolor           3.9
#> 61           5.0         2.0          3.5         1.0 versicolor           3.5
#> 62           5.9         3.0          4.2         1.5 versicolor           4.2
#> 63           6.0         2.2          4.0         1.0 versicolor           4.0
#> 64           6.1         2.9          4.7         1.4 versicolor           4.7
#> 65           5.6         2.9          3.6         1.3 versicolor           3.6
#> 66           6.7         3.1          4.4         1.4 versicolor           4.4
#> 67           5.6         3.0          4.5         1.5 versicolor           4.5
#> 68           5.8         2.7          4.1         1.0 versicolor           4.1
#> 69           6.2         2.2          4.5         1.5 versicolor           4.5
#> 70           5.6         2.5          3.9         1.1 versicolor           3.9
#> 71           5.9         3.2          4.8         1.8 versicolor           4.8
#> 72           6.1         2.8          4.0         1.3 versicolor           4.0
#> 73           6.3         2.5          4.9         1.5 versicolor           4.9
#> 74           6.1         2.8          4.7         1.2 versicolor           4.7
#> 75           6.4         2.9          4.3         1.3 versicolor           4.3
#> 76           6.6         3.0          4.4         1.4 versicolor           4.4
#> 77           6.8         2.8          4.8         1.4 versicolor           4.8
#> 78           6.7         3.0          5.0         1.7 versicolor           5.0
#> 79           6.0         2.9          4.5         1.5 versicolor           4.5
#> 80           5.7         2.6          3.5         1.0 versicolor           3.5
#> 81           5.5         2.4          3.8         1.1 versicolor           3.8
#> 82           5.5         2.4          3.7         1.0 versicolor           3.7
#> 83           5.8         2.7          3.9         1.2 versicolor           3.9
#> 84           6.0         2.7          5.1         1.6 versicolor           5.1
#> 85           5.4         3.0          4.5         1.5 versicolor           4.5
#> 86           6.0         3.4          4.5         1.6 versicolor           4.5
#> 87           6.7         3.1          4.7         1.5 versicolor           4.7
#> 88           6.3         2.3          4.4         1.3 versicolor           4.4
#> 89           5.6         3.0          4.1         1.3 versicolor           4.1
#> 90           5.5         2.5          4.0         1.3 versicolor           4.0
#> 91           5.5         2.6          4.4         1.2 versicolor           4.4
#> 92           6.1         3.0          4.6         1.4 versicolor           4.6
#> 93           5.8         2.6          4.0         1.2 versicolor           4.0
#> 94           5.0         2.3          3.3         1.0 versicolor           3.3
#> 95           5.6         2.7          4.2         1.3 versicolor           4.2
#> 96           5.7         3.0          4.2         1.2 versicolor           4.2
#> 97           5.7         2.9          4.2         1.3 versicolor           4.2
#> 98           6.2         2.9          4.3         1.3 versicolor           4.3
#> 99           5.1         2.5          3.0         1.1 versicolor           3.0
#> 100          5.7         2.8          4.1         1.3 versicolor           4.1
#> 101          6.3         3.3          6.0         2.5  virginica           6.0
#> 102          5.8         2.7          5.1         1.9  virginica           5.1
#> 103          7.1         3.0          5.9         2.1  virginica           5.9
#> 104          6.3         2.9          5.6         1.8  virginica           5.6
#> 105          6.5         3.0          5.8         2.2  virginica           5.8
#> 106          7.6         3.0          6.6         2.1  virginica           6.6
#> 107          4.9         2.5          4.5         1.7  virginica           4.5
#> 108          7.3         2.9          6.3         1.8  virginica           6.3
#> 109          6.7         2.5          5.8         1.8  virginica           5.8
#> 110          7.2         3.6          6.1         2.5  virginica           6.1
#> 111          6.5         3.2          5.1         2.0  virginica           5.1
#> 112          6.4         2.7          5.3         1.9  virginica           5.3
#> 113          6.8         3.0          5.5         2.1  virginica           5.5
#> 114          5.7         2.5          5.0         2.0  virginica           5.0
#> 115          5.8         2.8          5.1         2.4  virginica           5.1
#> 116          6.4         3.2          5.3         2.3  virginica           5.3
#> 117          6.5         3.0          5.5         1.8  virginica           5.5
#> 118          7.7         3.8          6.7         2.2  virginica           6.7
#> 119          7.7         2.6          6.9         2.3  virginica           6.9
#> 120          6.0         2.2          5.0         1.5  virginica           5.0
#> 121          6.9         3.2          5.7         2.3  virginica           5.7
#> 122          5.6         2.8          4.9         2.0  virginica           4.9
#> 123          7.7         2.8          6.7         2.0  virginica           6.7
#> 124          6.3         2.7          4.9         1.8  virginica           4.9
#> 125          6.7         3.3          5.7         2.1  virginica           5.7
#> 126          7.2         3.2          6.0         1.8  virginica           6.0
#> 127          6.2         2.8          4.8         1.8  virginica           4.8
#> 128          6.1         3.0          4.9         1.8  virginica           4.9
#> 129          6.4         2.8          5.6         2.1  virginica           5.6
#> 130          7.2         3.0          5.8         1.6  virginica           5.8
#> 131          7.4         2.8          6.1         1.9  virginica           6.1
#> 132          7.9         3.8          6.4         2.0  virginica           6.4
#> 133          6.4         2.8          5.6         2.2  virginica           5.6
#> 134          6.3         2.8          5.1         1.5  virginica           5.1
#> 135          6.1         2.6          5.6         1.4  virginica           5.6
#> 136          7.7         3.0          6.1         2.3  virginica           6.1
#> 137          6.3         3.4          5.6         2.4  virginica           5.6
#> 138          6.4         3.1          5.5         1.8  virginica           5.5
#> 139          6.0         3.0          4.8         1.8  virginica           4.8
#> 140          6.9         3.1          5.4         2.1  virginica           5.4
#> 141          6.7         3.1          5.6         2.4  virginica           5.6
#> 142          6.9         3.1          5.1         2.3  virginica           5.1
#> 143          5.8         2.7          5.1         1.9  virginica           5.1
#> 144          6.8         3.2          5.9         2.3  virginica           5.9
#> 145          6.7         3.3          5.7         2.5  virginica           5.7
#> 146          6.7         3.0          5.2         2.3  virginica           5.2
#> 147          6.3         2.5          5.0         1.9  virginica           5.0
#> 148          6.5         3.0          5.2         2.0  virginica           5.2
#> 149          6.2         3.4          5.4         2.3  virginica           5.4
#> 150          5.9         3.0          5.1         1.8  virginica           5.1
#>     Petal.Length2  sum
#> 1             1.4  2.8
#> 2             1.4  2.8
#> 3             1.3  2.6
#> 4             1.5  3.0
#> 5             1.4  2.8
#> 6             1.7  3.4
#> 7             1.4  2.8
#> 8             1.5  3.0
#> 9             1.4  2.8
#> 10            1.5  3.0
#> 11            1.5  3.0
#> 12            1.6  3.2
#> 13            1.4  2.8
#> 14            1.1  2.2
#> 15            1.2  2.4
#> 16            1.5  3.0
#> 17            1.3  2.6
#> 18            1.4  2.8
#> 19            1.7  3.4
#> 20            1.5  3.0
#> 21            1.7  3.4
#> 22            1.5  3.0
#> 23            1.0  2.0
#> 24            1.7  3.4
#> 25            1.9  3.8
#> 26            1.6  3.2
#> 27            1.6  3.2
#> 28            1.5  3.0
#> 29            1.4  2.8
#> 30            1.6  3.2
#> 31            1.6  3.2
#> 32            1.5  3.0
#> 33            1.5  3.0
#> 34            1.4  2.8
#> 35            1.5  3.0
#> 36            1.2  2.4
#> 37            1.3  2.6
#> 38            1.4  2.8
#> 39            1.3  2.6
#> 40            1.5  3.0
#> 41            1.3  2.6
#> 42            1.3  2.6
#> 43            1.3  2.6
#> 44            1.6  3.2
#> 45            1.9  3.8
#> 46            1.4  2.8
#> 47            1.6  3.2
#> 48            1.4  2.8
#> 49            1.5  3.0
#> 50            1.4  2.8
#> 51            4.7  9.4
#> 52            4.5  9.0
#> 53            4.9  9.8
#> 54            4.0  8.0
#> 55            4.6  9.2
#> 56            4.5  9.0
#> 57            4.7  9.4
#> 58            3.3  6.6
#> 59            4.6  9.2
#> 60            3.9  7.8
#> 61            3.5  7.0
#> 62            4.2  8.4
#> 63            4.0  8.0
#> 64            4.7  9.4
#> 65            3.6  7.2
#> 66            4.4  8.8
#> 67            4.5  9.0
#> 68            4.1  8.2
#> 69            4.5  9.0
#> 70            3.9  7.8
#> 71            4.8  9.6
#> 72            4.0  8.0
#> 73            4.9  9.8
#> 74            4.7  9.4
#> 75            4.3  8.6
#> 76            4.4  8.8
#> 77            4.8  9.6
#> 78            5.0 10.0
#> 79            4.5  9.0
#> 80            3.5  7.0
#> 81            3.8  7.6
#> 82            3.7  7.4
#> 83            3.9  7.8
#> 84            5.1 10.2
#> 85            4.5  9.0
#> 86            4.5  9.0
#> 87            4.7  9.4
#> 88            4.4  8.8
#> 89            4.1  8.2
#> 90            4.0  8.0
#> 91            4.4  8.8
#> 92            4.6  9.2
#> 93            4.0  8.0
#> 94            3.3  6.6
#> 95            4.2  8.4
#> 96            4.2  8.4
#> 97            4.2  8.4
#> 98            4.3  8.6
#> 99            3.0  6.0
#> 100           4.1  8.2
#> 101           6.0 12.0
#> 102           5.1 10.2
#> 103           5.9 11.8
#> 104           5.6 11.2
#> 105           5.8 11.6
#> 106           6.6 13.2
#> 107           4.5  9.0
#> 108           6.3 12.6
#> 109           5.8 11.6
#> 110           6.1 12.2
#> 111           5.1 10.2
#> 112           5.3 10.6
#> 113           5.5 11.0
#> 114           5.0 10.0
#> 115           5.1 10.2
#> 116           5.3 10.6
#> 117           5.5 11.0
#> 118           6.7 13.4
#> 119           6.9 13.8
#> 120           5.0 10.0
#> 121           5.7 11.4
#> 122           4.9  9.8
#> 123           6.7 13.4
#> 124           4.9  9.8
#> 125           5.7 11.4
#> 126           6.0 12.0
#> 127           4.8  9.6
#> 128           4.9  9.8
#> 129           5.6 11.2
#> 130           5.8 11.6
#> 131           6.1 12.2
#> 132           6.4 12.8
#> 133           5.6 11.2
#> 134           5.1 10.2
#> 135           5.6 11.2
#> 136           6.1 12.2
#> 137           5.6 11.2
#> 138           5.5 11.0
#> 139           4.8  9.6
#> 140           5.4 10.8
#> 141           5.6 11.2
#> 142           5.1 10.2
#> 143           5.1 10.2
#> 144           5.9 11.8
#> 145           5.7 11.4
#> 146           5.2 10.4
#> 147           5.0 10.0
#> 148           5.2 10.4
#> 149           5.4 10.8
#> 150           5.1 10.2