Lookup the Details of Parameters Required by a Model

This function simplifies the process of checking which parameters a given idmodelr model depends on. It is effectively an interface to parameter_details via model_details. As fuzzy matching has been used it can also given information of the parameter requirements of a subset of the available models.

required_parameters(model = NULL)

Arguments

model
A character string containing the name of the model of interest. Defaults to NULL.

Value

A dataframe extracted from parameter_details containing the details of the parameters required by the model of interest.

Examples

## Check the parameters required by the "SIR_ode" model
required_parameters("SIR_ode")

#> # A tibble: 2 x 6
#>   parameter parameter_family description   type  risk_stratified non_exponential
#>   <chr>     <chr>            <chr>         <chr> <chr>           <chr>          
#> 1 beta      transmission     Transmission… rate  no              no             
#> 2 tau       recovery         Recovery rat… rate  no              no             

## Use fizzy matching to look at paramters for all SIR models
required_parameters("SIR")

#> # A tibble: 5 x 6
#>   parameter parameter_family description   type  risk_stratified non_exponential
#>   <chr>     <chr>            <chr>         <chr> <chr>           <chr>          
#> 1 beta      transmission     Transmission… rate  no              no             
#> 2 tau       recovery         Recovery rat… rate  no              no             
#> 3 lambda    vaccination      The effectiv… prob… no              no             
#> 4 alpha     vaccination      The coverage… prop… no              no             
#> 5 mu        demographics     The natural … rate  no              no