Sum a Stratified Variable

Sum a Stratified Variable

summarise_strat_var(df, vars, strat = NULL, new_var = "sum")

Arguments

df
A dataframe of model output.

vars
A character vector containing the unstratified variables to summarise

strat
The number of stratifications present in the data set

new_var
The name of the summarised variable

Value

Returns the original dataframe with an additional summarised variable

See also

summarise_var_by_strat

Examples

df <- dplyr::mutate(iris, Petal.Length1 = Petal.Length, Petal.Length2 = Petal.Length)
df <- tibble::as_tibble(df)

summarise_strat_var(df, vars = c("Petal.Length"), strat = 2, new_var = "sum")

#> # A tibble: 150 x 8
#>      sum Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>    <dbl>        <dbl>       <dbl>        <dbl>       <dbl>  <fctr>
#>  1   2.8          5.1         3.5          1.4         0.2  setosa
#>  2   2.8          4.9         3.0          1.4         0.2  setosa
#>  3   2.6          4.7         3.2          1.3         0.2  setosa
#>  4   3.0          4.6         3.1          1.5         0.2  setosa
#>  5   2.8          5.0         3.6          1.4         0.2  setosa
#>  6   3.4          5.4         3.9          1.7         0.4  setosa
#>  7   2.8          4.6         3.4          1.4         0.3  setosa
#>  8   3.0          5.0         3.4          1.5         0.2  setosa
#>  9   2.8          4.4         2.9          1.4         0.2  setosa
#> 10   3.0          4.9         3.1          1.5         0.1  setosa
#> # ... with 140 more rows, and 2 more variables: Petal.Length1 <dbl>,
#> #   Petal.Length2 <dbl>